[*] Fix CI: SVE directory scope, endian.h on Windows, format/overflow on ARM32
- ECDSA SVE: move set_source_files_properties to parent crypto/CMakeLists.txt
  (CMake directory scope issue - properties set in subdirectory were invisible)
- Windows: guard endian.h include in migrate_mdbx.c with #ifndef DAP_OS_WINDOWS
- ARM32: fix %lu format for uint64_t in test_globaldb_migrate.c (use %llu cast)
- ARM32: guard 52-bit limb tests in test_field_mul.c with #ifdef ECDSA_FIELD_52BIT

+++ tests/unit/global-db/test_globaldb_migrate.c
@@ -205,7 +205,7 @@ static void test_migrate_verify_data(void)
         dap_global_db_storage_init(TEST_DEST_DIR);
         
         uint64_t count = dap_global_db_storage_group_count("test_group", false);
-        dap_test_msg("test_group has %lu records", count);
+        dap_test_msg("test_group has %llu records", (unsigned long long)count);
         
         dap_global_db_storage_deinit();
     }
